Add donated buffer multi-functions edge case unit test.
The underlying issue was fixed in IPU XLA TensorFlow commit `f241ed116b091fa4ddfa4c20044fdcea6559172b`:
https://github.com/graphcore-research/tensorflow-jax-experimental/commit/f241ed116b091fa4ddfa4c20044fdcea6559172b

This PR is only adding additional testing at the JAX level to make sure
an exception is raised instead of executing the jitted function and
returning the wrong result.

In short: this edge case only appears when
an SRAM donated buffer was used in another function, and the IPU PjRt
client was not checking if it would become invalid (as a new engine is
loaded on the IPU).